MainWindow.java: updated the application's icon (dfscan3.png).
FilterBuilderWindow.java renamed to FilterBuilderDialog.java

Report.java: added an attribute to the result reports that reflect the
name of the detection profile used to generate the report.

ProgressPanel.java, ResultsPanel.java: these classes now add a tab
instance to their parent JTabbedPane, updated the tabs' titles, did some
member variable refactoring.

TabbedPanel.java: removed the abstract method declaration in favor of an
overridable member method so that the default action when selecting a
tab's close button is to call the abstract method closePanel() defined
in the extending classes.